Update tools version and fixed permission bug
If an app using a permission is installed before the app
providing the permission has been installed, a security exception
is thrown and the requiring app will not work until a restart of
the device. This is fixed by providing the permission in both the
provider and the consumer of the permission.